Reorganized utils
