--- apps/shade/src/utils/debounce.ts
@@ -1,24 +1,0 @@
-export function debounce<T extends unknown[]>(func: (...args: T) => void, wait: number, immediate: boolean = false): (...args: T) => void {
-    let timeoutId: ReturnType<typeof setTimeout> | null;
-
-    return function (this: unknown, ...args: T): void {
-        const later = () => {
-            timeoutId = null;
-            if (!immediate) {
-                func.apply(this, args);
-            }
-        };
-
-        const callNow = immediate && !timeoutId;
-
-        if (timeoutId) {
-            clearTimeout(timeoutId);
-        }
-
-        timeoutId = setTimeout(later, wait);
-
-        if (callNow) {
-            func.apply(this, args);
-        }
-    };
-}
