--- apps/shade/src/utils/formatText.ts
@@ -1,6 +1,0 @@
-// Helper to convert kebab-case to PascalCase with numbers
-export const kebabToPascalCase = (str: string): string => {
-    const processed = str
-        .replace(/[-_]([a-z0-9])/gi, (_, char) => char.toUpperCase());
-    return processed.charAt(0).toUpperCase() + processed.slice(1);
-};
